event_bus関連のテスト追加
